refactor(web): move useThemeColor to src/theme
