+++ packages/web-reporter-ui/index.tsx
@@ -1,3 +1,3 @@
 export { IterationsReporterView, ReporterView } from "./ReporterView";
 export { PageBackground } from "./components/PageBackground";
-export { useSetThemeAtRandom } from "./useThemeColor";
+export { useSetThemeAtRandom } from "./src/theme/useThemeColor";
